# scripts/openapi.json
{
  "components": {
    "schemas": {
      "AdminCreditGrantRequest": {
        "properties": {
          "amount": {
            "exclusiveMinimum": 0.0,
            "title": "Amount",
            "type": "integer"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "required": [
          "amount"
        ],
        "title": "AdminCreditGrantRequest",
        "type": "object"
      },
      "AdminCreditGrantResponse": {
        "properties": {
          "balance": {
            "title": "Balance",
            "type": "integer"
          },
          "granted": {
            "title": "Granted",
            "type": "integer"
          },
          "userId": {
            "title": "Userid",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "balance",
          "granted"
        ],
        "title": "AdminCreditGrantResponse",
        "type": "object"
      },
      "AdminOrderItemListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminOrderItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "AdminOrderItemListResponse",
        "type": "object"
      },
      "AdminOrderItemResponse": {
        "properties": {
          "createdAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          },
          "fulfillmentStatus": {
            "enum": [
              "paid",
              "preparing",
              "shipped",
              "delivered"
            ],
            "title": "Fulfillmentstatus",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lineTotalCredits": {
            "title": "Linetotalcredits",
            "type": "integer"
          },
          "orderId": {
            "title": "Orderid",
            "type": "string"
          },
          "productId": {
            "title": "Productid",
            "type": "string"
          },
          "productTitle": {
            "title": "Producttitle",
            "type": "string"
          },
          "qty": {
            "title": "Qty",
            "type": "integer"
          },
          "sellerType": {
            "enum": [
              "platform",
              "merchant"
            ],
            "title": "Sellertype",
            "type": "string"
          },
          "shopName": {
            "title": "Shopname",
            "type": "string"
          },
          "unitPriceCredits": {
            "title": "Unitpricecredits",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "orderId",
          "productId",
          "productTitle",
          "qty",
          "unitPriceCredits",
          "lineTotalCredits",
          "fulfillmentStatus",
          "shopName",
          "sellerType"
        ],
        "title": "AdminOrderItemResponse",
        "type": "object"
      },
      "AdminSellerItem": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "sellerType": {
            "title": "Sellertype",
            "type": "string"
          },
          "shopName": {
            "title": "Shopname",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "userEmail": {
            "title": "Useremail",
            "type": "string"
          },
          "userId": {
            "title": "Userid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "userEmail",
          "shopName",
          "slug",
          "status",
          "sellerType",
          "createdAt"
        ],
        "title": "AdminSellerItem",
        "type": "object"
      },
      "AdminSellerListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminSellerItem"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "AdminSellerListResponse",
        "type": "object"
      },
      "AdminStatsResponse": {
        "properties": {
          "orderCount": {
            "title": "Ordercount",
            "type": "integer"
          },
          "pendingSellerCount": {
            "title": "Pendingsellercount",
            "type": "integer"
          },
          "productCount": {
            "title": "Productcount",
            "type": "integer"
          },
          "sellerCount": {
            "title": "Sellercount",
            "type": "integer"
          },
          "userCount": {
            "title": "Usercount",
            "type": "integer"
          }
        },
        "required": [
          "userCount",
          "productCount",
          "orderCount",
          "sellerCount",
          "pendingSellerCount"
        ],
        "title": "AdminStatsResponse",
        "type": "object"
      },
      "AdminUserItem": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "displayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "isAdmin": {
            "title": "Isadmin",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "isAdmin",
          "createdAt"
        ],
        "title": "AdminUserItem",
        "type": "object"
      },
      "AdminUserListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminUserItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "AdminUserListResponse",
        "type": "object"
      },
...
      "DbResetRequest": {
        "properties": {
          "confirm": {
            "title": "Confirm",
            "type": "string"
          },
          "mode": {
            "default": "seed",
            "enum": [
              "seed",
              "truncate_all",
              "truncate_except_users"
            ],
            "title": "Mode",
            "type": "string"
          }
        },
        "required": [
          "confirm"
        ],
        "title": "DbResetRequest",
        "type": "object"
      },
      "DbResetResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "mode": {
            "enum": [
              "seed",
              "truncate_all",
              "truncate_except_users"
            ],
            "title": "Mode",
            "type": "string"
          }
        },
        "required": [
          "mode",
          "message"
        ],
        "title": "DbResetResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SellerOrderItemStatusUpdate": {
        "properties": {
          "fulfillmentStatus": {
            "enum": [
              "paid",
              "preparing",
              "shipped",
              "delivered"
            ],
            "title": "Fulfillmentstatus",
            "type": "string"
          }
        },
        "required": [
          "fulfillmentStatus"
        ],
        "title": "SellerOrderItemStatusUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Shopping mall backend skeleton",
    "title": "Shopping Mall API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/db/reset": {
      "post": {
        "operationId": "reset_database_admin_db_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DbResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DbResetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reset Database",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/orders": {
      "get": {
        "operationId": "list_admin_orders_admin_orders_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderItemListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Admin Orders",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/orders/items/{item_id}/status": {
      "patch": {
        "operationId": "admin_update_order_item_status_admin_orders_items__item_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SellerOrderItemStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Admin Update Order Item Status",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/sellers": {
      "get": {
        "operationId": "list_sellers_admin_sellers_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSellerListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Sellers",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/sellers/{seller_id}/approve": {
      "post": {
        "operationId": "approve_seller_endpoint_admin_sellers__seller_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "seller_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Seller Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSellerItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve Seller Endpoint",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/sellers/{seller_id}/suspend": {
      "post": {
        "operationId": "suspend_seller_endpoint_admin_sellers__seller_id__suspend_post",
        "parameters": [
          {
            "in": "path",
            "name": "seller_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Seller Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSellerItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Suspend Seller Endpoint",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/stats": {
      "get": {
        "operationId": "admin_stats_admin_stats_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminStatsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Admin Stats",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users": {
      "get": {
        "operationId": "list_users_admin_users_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Users",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/{user_id}/credits": {
      "post": {
        "operationId": "grant_user_credits_admin_users__user_id__credits_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCreditGrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCreditGrantResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Grant User Credits",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/{user_id}/promote": {
      "post": {
        "operationId": "promote_user_admin_users__user_id__promote_post",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Promote User",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}